+++ src/main/java/org/bytedeco/javacpp/BytePointer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2015 Samuel Audet
+ * Copyright (C) 2011-2016 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -120,9 +120,9 @@ public class BytePointer extends Pointer {
     public byte[] getStringBytes() {
         long size = limit - position;
         if (size <= 0) {
-            size = (int)strlen(this);
+            size = strlen(this);
         }
-        byte[] array = new byte[(int) size];
+        byte[] array = new byte[(int)Math.min(size, Integer.MAX_VALUE)];
         get(array);
         return array;
     }