+++ src/main/java/org/bytedeco/javacpp/annotation/ArrayAllocator.java
@@ -19,7 +19,7 @@ import org.bytedeco.javacpp.tools.Generator;
  * with {@code NativeDeallocator}, based on the {@code delete[]} operator, if
  * not additionally annotated with {@link NoDeallocator}.
  *
- * @see Pointer#init(long, int, long, long)
+ * @see Pointer#init(long, long, long, long)
  * @see Generator
  *
  * @author Samuel Audet