+++ src/main/java/org/bytedeco/javacpp/indexer/FloatIndexer.java
@@ -59,7 +59,7 @@ public abstract class FloatIndexer extends Indexer {
             return new FloatBufferIndexer(pointer.asBuffer(), sizes, strides);
         } else {
             final long position = pointer.position();
-            float[] array = new float[(int) (pointer.limit() - position)];
+            float[] array = new float[(int)Math.min(pointer.limit() - position, Integer.MAX_VALUE)];
             pointer.get(array);
             return new FloatArrayIndexer(array, sizes, strides) {
                 @Override public void release() {